+++ src/Serializers/RBX.php
@@ -150,7 +150,7 @@ class RBX implements Serializer
         $header = JSON::decode($header);
 
         if ($version <= 0 or $version > 2) {
-            throw new RuntimeException("Incompatible with RBX version $version.");
+            throw new RuntimeException("Incompatible with RBX $version format.");
         }
 
         if (strlen($payload) !== $header['data']['length']) {