+++ src/Transformers/RegexFilter.php
@@ -34,13 +34,6 @@ class RegexFilter implements Transformer
     public const EMAIL = '/[a-z0-9_\-\+\.]+@[a-z0-9\-]+\.([a-z]{2,4})(?:\.[a-z]{2})?/i';
 
     /**
-     * A pattern to match unicode emojis.
-     *
-     * @var string
-     */
-    public const EMOJIS = '/[\x{1F300}-\x{1F5FF}\x{1F900}-\x{1F9FF}\x{1F600}-\x{1F64F}\x{1F680}-\x{1F6FF}\x{2600}-\x{26FF}\x{2700}-\x{27BF}]/u';
-
-    /**
      * The default URL matching pattern.
      *
      * @var literal-string