+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     - Added validation interval parameter to MLPs and GBM Learners
     - Removed output layer L2 Penalty parameter from MLP Learners
     - Remove Network interface
+    - RBX Serializer only tracks major library version number
     
 - 2.4.1
     - Sentence Tokenizer fix Arabic and Farsi language support